Refactor: 기존 model.rs 파일을 없애고 Rust Module & 디렉토리로 데이터 모델 구현 구조 변경
